requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.7
djangorestframework==3.15.2
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jiter==0.9.0
multidict==6.2.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
openai==0.28.0
packaging==24.2
pandas==2.2.3
propcache==0.3.0
psycopg2==2.9.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
